Conflicts:
	core/src/main/scala/spark/Executor.scala
	core/src/main/scala/spark/RDD.scala

+++ core/src/main/scala/spark/Job.scala
@@ -1,13 +1,14 @@
 package spark
 
-import mesos._
+import org.apache.mesos._
+import org.apache.mesos.Protos._
 
 /**
  * Class representing a parallel job in MesosScheduler. Schedules the
  * job by implementing various callbacks.
  */
 abstract class Job(jobId: Int) {
-  def slaveOffer(s: SlaveOffer, availableCpus: Int, availableMem: Int)
+  def slaveOffer(s: SlaveOffer, availableCpus: Double, availableMem: Double)
     : Option[TaskDescription]
 
   def statusUpdate(t: TaskStatus): Unit
